# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HeaSys API Documentation",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local"
        }
    ],
    "paths": {
        "/api/appointments": {
            "get": {
                "tags": [
                    "Appointment"
                ],
                "summary": "List all appointments",
                "description": "Fetch a list of all appointments",
                "operationId": "812c4f15927f3d0e3e959538a6f1104b",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AppointmentResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Appointment"
                ],
                "summary": "Create new appointment",
                "description": "Create new appointment",
                "operationId": "2b481df28ec53d2afb2ec7c9a3a77846",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreAppointmentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AppointmentResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "patient_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The selected patient is invalid."
                                                    }
                                                },
                                                "diagnose_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The selected diagnose is invalid."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/appointments/{id}": {
            "get": {
                "tags": [
                    "Appointment"
                ],
                "summary": "Show specific appointment",
                "description": "Show specific appointment",
                "operationId": "cca217dc71eae262de1fe09f314a4948",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AppointmentResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Appointment not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Appointment"
                ],
                "summary": "Update specific appointment",
                "description": "Update specific appointment",
                "operationId": "6baab7054cff6982e42d5346940917bc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAppointmentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AppointmentResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Appointment not found or update failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "patient_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The selected patient is invalid."
                                                    }
                                                },
                                                "diagnose_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The selected diagnose is invalid."
                                                    }
                                                },
                                                "status": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The selected status is invalid."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/appointments/{id}/services/{service_id}/status": {
            "patch": {
                "tags": [
                    "Appointment"
                ],
                "summary": "Update the service status of a specific appointment",
                "description": "Update the service status of a specific appointment",
                "operationId": "7c1cf1e28b5f1fbb455f52da4694db05",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "service_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateStatusRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AppointmentResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Appointment not found or update failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "status": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The selected status is invalid."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreAppointmentRequest": {
                "title": "Store Appointment Request",
                "required": [
                    "patient_id",
                    "diagnose_id"
                ],
                "properties": {
                    "patient_id": {
                        "description": "ID of the patient",
                        "type": "integer",
                        "example": 1
                    },
                    "diagnose_id": {
                        "description": "ID of the diagnose",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "UpdateAppointmentRequest": {
                "title": "Update Appointment Request",
                "required": [
                    "patient_id",
                    "diagnose_id",
                    "status"
                ],
                "properties": {
                    "patient_id": {
                        "description": "ID of the patient",
                        "type": "integer",
                        "example": 1
                    },
                    "diagnose_id": {
                        "description": "ID of the diagnose",
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "description": "1 = process, 0 = success",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "UpdateStatusRequest": {
                "title": "Update Status Request",
                "required": [
                    "status"
                ],
                "properties": {
                    "status": {
                        "description": "1 = process, 0 = success",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "AppointmentResource": {
                "title": "Appointment Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "description": "1 = process, 0 = success",
                        "type": "integer",
                        "example": 1
                    },
                    "patient": {
                        "$ref": "#/components/schemas/PatientResource"
                    },
                    "diagnose": {
                        "$ref": "#/components/schemas/DiagnoseResource"
                    },
                    "checkups": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CheckupProgressResource"
                        }
                    }
                },
                "type": "object"
            },
            "CheckupProgressResource": {
                "title": "Checkup Progress Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "description": "1 = process, 0 = success",
                        "type": "integer",
                        "example": 1
                    },
                    "service": {
                        "$ref": "#/components/schemas/ServiceResource"
                    }
                },
                "type": "object"
            },
            "DiagnoseResource": {
                "title": "Diagnose Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Diagnosa Example"
                    }
                },
                "type": "object"
            },
            "PatientResource": {
                "title": "Patient Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Patient Example"
                    }
                },
                "type": "object"
            },
            "ServiceResource": {
                "title": "Service Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Service Example"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}